# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "description": "API Docs for Inventory App",
    "version": "1.0",
    "title": "Inventory API"
  },
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "servers": [
    {
      "url": "localhost:8080/api/v1"
    }
  ],
  "paths": {
...
    "/lists": {
      "post": {
        "summary": "Create a Shopping List",
        "description": "Create a new Shopping List for the requesting user. Users with free accounts can create up to one Shopping List at a time. Premium users can create unlimited Shopping Lists.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ListCreateBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/SuccessResponse"
          },
          "400": {
            "$ref": "#/components/responses/BadFormatResponse"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedResponse"
          }
        }
      },
      "get": {
        "summary": "Get all Shopping Lists",
        "description": "Get a list of all Shopping Lists. Free users can view Shopping Lists which they own. Premium users can view Shopping Lists which they own or are a member of. Lists are sorted by last updated time. Lists can be filtered by owner or member.",
        "parameters": [
          {
            "$ref": "#/components/parameters/OwnerQuery"
          },
          {
            "$ref": "#/components/parameters/MemberQuery"
          },
          {
            "in": "query",
            "name": "sort",
            "description": "Field to sort Shopping Lists by",
            "schema": {
              "type": "string",
              "default": "updatedAt",
              "enum": ["name", "createdAt", "updatedAt"]
            }
          },
          {
            "$ref": "#/components/parameters/OrderQuery"
          },
          {
            "$ref": "#/components/parameters/LimitQuery"
          },
          {
            "$ref": "#/components/parameters/OffsetQuery"
          }
        ],
        "responses": {
          "200": {
            "description": "A list of Shopping Lists",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ShoppingListBasic"
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedResponse"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundResponse"
          }
        }
      }
    },
    "/lists/{id}": {
      "get": {
        "summary": "Get a Shopping List by ID",
        "description": "Get a Shopping List by ID. Will only return a Shopping List if a user is authorized to view it. Free users can view Shopping Lists which they own. Premium users can view Shopping Lists which they own or are a member of.",
        "parameters": [
          {
            "$ref": "#/components/parameters/ListIdParam"
          }
        ],
        "responses": {
          "200": {
            "description": "The requested Shopping List object",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShoppingList"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedResponse"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundResponse"
          }
        }
      },
      "put": {
        "summary": "Update a Shopping List",
        "description": "Update the Shopping List with the given ID. Updatable fields include name and notes. Users with free accounts can update their default Shopping List. Premium users can update Shopping Lists which they own or are a member of, provided that the owner has an active subscription.",
        "parameters": [
          {
            "$ref": "#/components/parameters/ListIdParam"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ListUpdateBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/SuccessResponse"
          },
          "400": {
            "$ref": "#/components/responses/BadFormatResponse"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedResponse"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundResponse"
          }
        }
      },
      "delete": {
        "summary": "Delete a Shopping List",
        "description": "Delete a Shopping List. Users can delete Shopping Lists which they own, regardless of subscription status.",
        "parameters": [
          {
            "$ref": "#/components/parameters/ListIdParam"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/SuccessResponse"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedResponse"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundResponse"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "Name": {
        "type": "string",
        "description": "The user's name or username",
        "example": "jjfino"
      },
      "PhotoUrl": {
        "type": "string",
        "description": "The url to the User's profile photo",
        "example": "www.google.com"
      },
...
      "IsSubscribed": {
        "type": "boolean",
        "description": "Whether this user's subscription is active at the time of retrieval"
      },
...
      "ShoppingList": {
        "type": "object",
        "required": ["name", "owner", "items", "members", "lastUpdatedBy"],
        "properties": {
          "name": {
            "$ref": "#/components/schemas/ListName"
          },
          "notes": {
            "$ref": "#/components/schemas/ListNotes"
          },
          "owner": {
            "$ref": "#/components/schemas/BaseUserWithExpiry"
          },
          "members": {
            "type": "array",
            "description": "The members of the Shopping List",
            "items": {
              "$ref": "#/components/schemas/BaseUser"
            }
          },
          "items": {
            "type": "array",
            "description": "The items in the Shopping List",
            "items": {
              "$ref": "#/components/schemas/ListItem"
            }
          },
          "lastUpdatedBy": {
            "$ref": "#/components/schemas/BaseUser"
          }
        }
      },
      "ShoppingListBasic": {
        "type": "object",
        "required": ["name", "owner", "items", "members", "lastUpdatedBy"],
        "properties": {
          "name": {
            "$ref": "#/components/schemas/ListName"
          },
          "notes": {
            "$ref": "#/components/schemas/ListNotes"
          },
          "owner": {
            "$ref": "#/components/schemas/BaseUserWithExpiry"
          },
          "members": {
            "type": "array",
            "description": "The members of the Shopping List",
            "items": {
              "$ref": "#/components/schemas/BaseUser"
            }
          },
          "items": {
            "type": "integer",
            "description": "The number of items in the Shopping List"
          },
          "lastUpdatedBy": {
            "$ref": "#/components/schemas/BaseUser"
          }
        }
      },
      "ListCreateBody": {
        "type": "object",
        "required": ["name"],
        "properties": {
          "name": {
            "$ref": "#/components/schemas/ListItemName"
          },
          "notes": {
            "$ref": "#/components/schemas/ListItemNotes"
          }
        }
      },
      "ListUpdateBody": {
        "type": "object",
        "properties": {
          "name": {
            "$ref": "#/components/schemas/ListItemName"
          },
          "notes": {
            "$ref": "#/components/schemas/ListItemNotes"
          },
          "newOrder": {
            "description": "The new order of Items in the Shopping List",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ListItemOrder"
            }
          }
        }
      },
      "ListItem": {
        "type": "object",
        "required": ["name", "checked"],
        "properties": {
          "name": {
            "$ref": "#/components/schemas/ListItemName"
          },
          "notes": {
            "$ref": "#/components/schemas/ListItemNotes"
          },
          "checked": {
            "$ref": "#/components/schemas/ListItemChecked"
          },
          "pos": {
            "$ref": "#/components/schemas/ListItemOrder"
          }
        }
      },
...
      "ObjectID": {
        "type": "string",
        "pattern": "^[a-z0-9]{24}$",
        "description": "The object's ID"
      },
...
      "ListName": {
        "type": "string",
        "description": "The name of the Shopping List",
        "example": "Whole Foods Shopping List",
        "maxLength": 100
      },
      "ListNotes": {
        "type": "string",
        "description": "Any notes about the Shopping List",
        "example": "List for this weekend",
        "maxLength": 300
      },
      "ListItemName": {
        "type": "string",
        "description": "The name of the item",
        "example": "Eggs",
        "maxLength": 100
      },
      "ListItemNotes": {
        "type": "string",
        "description": "Any notes about the item",
        "example": "2% milk",
        "maxLength": 300
      },
      "ListItemChecked": {
        "type": "boolean",
        "description": "Whether this item is checked off in the list"
      },
      "ListItemOrder": {
        "type": "integer",
        "description": "The order of this Item in the Shopping List",
        "minimum": 0
      },
...
      "BaseUser": {
        "type": "object",
        "required": ["id", "name", "photoUrl"],
        "properties": {
          "id": {
            "$ref": "#/components/schemas/ObjectID"
          },
          "name": {
            "$ref": "#/components/schemas/Name"
          },
          "photoUrl": {
            "$ref": "#/components/schemas/PhotoUrl"
          }
        }
      },
      "BaseUserWithExpiry": {
        "type": "object",
        "required": ["id", "name", "photoUrl"],
        "properties": {
          "id": {
            "$ref": "#/components/schemas/ObjectID"
          },
          "name": {
            "$ref": "#/components/schemas/Name"
          },
          "photoUrl": {
            "$ref": "#/components/schemas/PhotoUrl"
          },
          "isSubscribed": {
            "$ref": "#/components/schemas/IsSubscribed"
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "required": ["name", "statusCode", "message", "details"],
        "properties": {
          "name": {
            "description": "The name of the error",
            "type": "string",
            "example": "database-error"
          },
          "message": {
            "description": "A message describing the error",
            "type": "string",
            "example": "Record not found"
          },
          "statusCode": {
            "description": "The status code of the error",
            "type": "string",
            "example": "404"
          },
          "details": {
            "description": "Error details",
            "type": "string",
            "example": "This resource was not found or you do not have permission to access it."
          }
        }
      }
    },
    "parameters": {
      "OwnerQuery": {
        "in": "query",
        "name": "owner",
        "description": "The id of the owner of this resource",
        "schema": {
          "$ref": "#/components/schemas/ObjectID"
        }
      },
      "MemberQuery": {
        "in": "query",
        "name": "member",
        "description": "The id of the member of this resource",
        "schema": {
          "$ref": "#/components/schemas/ObjectID"
        }
      },
      "LimitQuery": {
        "in": "query",
        "name": "limit",
        "required": false,
        "description": "The maximum number of results to return",
        "schema": {
          "type": "integer",
          "default": 30,
          "minimum": 0,
          "maximum": 50
        }
      },
...
      "OffsetQuery": {
        "in": "query",
        "name": "offset",
        "required": false,
        "description": "Offset of results to return",
        "schema": {
          "type": "integer",
          "default": "0",
          "minimum": 0
        }
      },
...
      "OrderQuery": {
        "in": "query",
        "name": "order",
        "required": false,
        "description": "Whether to sort in descending order",
        "schema": {
          "type": "string",
          "enum": ["asc", "desc"],
          "default": "asc"
        }
      },
...
      "ListIdParam": {
        "in": "path",
        "name": "id",
        "required": true,
        "description": "The ID of the Shopping List",
        "schema": {
          "$ref": "#/components/schemas/ObjectID"
        }
      },
...
    },
    "responses": {
      "SuccessResponse": {
        "description": "A success message",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": ["message"],
              "properties": {
                "message": {
                  "description": "A success message",
                  "type": "string",
                  "example": "Update successful"
                }
              }
            }
          }
        }
      },
      "NotFoundResponse": {
        "description": "The requested resource was not found or the user's account level does not have permission to access it.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "UnauthorizedResponse": {
        "description": "User is unauthorized. This can be due to invalid credentials or not having a premium account.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "BadFormatResponse": {
        "description": "The request was formed badly. Perhaps required fields are missing or invalid in the request body.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      }
    }
  }
}